+++ CHANGES.md
@@ -5,7 +5,7 @@
   - Patch ISMRMRD `CMAKE_INSTALL_RPATH_USE_LINK_PATH` (#991)
 - Updated versions:
   - SIRF-Contribs: v3.9.1
-  - SIRF: 4198e7f17ef17ff8b1913e20d1db3e7c9383d4da (2026-03-12)
+  - SIRF: 6e429c2d1862057086720c543d220730feb3eecb (2026-05-12)
   - CCPi-Regularisation-Toolkit, astra-toolbox, TomoPhantom: moved to conda
 
 ## v3.9.0